Breaking load and verify out into privatekey package
