+++ docker-compose.yml
@@ -39,8 +39,6 @@ services:
 
   bmysql:
     image: mariadb:10.5
-    ports:
-      - 3306:3306 # DB
     networks:
       bluenet:
         aliases:
